% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antlers.R
\docType{data}
\name{antlers}
\alias{antlers}
\title{Simulated data on the effects of diet on deer antlers.}
\format{A data frame
\describe{
  \item{diet}{Experimental diet during both spring and summer.  Hi.Hi = High quality diet in both season.  Hi.Lo = high quality diet in spring but poor in summer.  }
  \item{mass}{Antler mass in grams}
  \item{circum}{basal circumference in mm}
  \item{beam}{Length of main beam of antler in mm}
  \item{spread}{spread between points in mm}
  \item{diet.name}{further details on diet treatment}
}}
\usage{
antlers
}
\description{
Data simulated based on summary information in Asleson et al. (1997)
}
\references{
Asleson et al. 1997.  Effects of seasonal protein restriction on antlerogenesis and body mass in adult male white-tailed deer.  Journal of Wildlife Management 61.
}
\keyword{datasets}
