% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOU_species_codes.R
\docType{data}
\name{AOU_species_codes}
\alias{AOU_species_codes}
\title{American Ornithological Union (AOU) species codes for North-American birds}
\format{A data frame...
\describe{
  \item{spp.num}{AOU species codes}
  \item{alpha.code}{Bird Conservation Region}
  \item{name}{Common name of species.}
  \item{spp}{Latin binomial genus and species}
  \item{order}{code for taxonomic order}
 }}
\usage{
AOU_species_codes
}
\description{
Some records from offical list have been removed because they don't have stable names, eg., they are a defined species.
}
\keyword{datasets}
