% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBS_PA.R
\docType{data}
\name{BBS_PA}
\alias{BBS_PA}
\title{Pennsylvania Breeding Survey Data (BBS) from the USGS.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{countrynum}{Numeric code indicating country.  USA = 840.}
  \item{statenum}{Numeric code indicating state.  PA = 72.}
  \item{Route}{numeric code indicating router number}
  \item{RPID}{Run Protocol Identification Number.  See RunProtocolID.txt file in USGS documentation}
  \item{Year}{Yeardata collected.}
  \item{Aou}{numeric species code from American Ornithological Union}
  \item{StopTotal}{Total number of stops out of 50 on which the species was recorded.  This is not a count of birds, but the number of stops that had a particular bird!  The min value is 1, and the max value is 50}
  \item{SpeciesTotal}{Total number of stops out of 50 on which the species was recorded.  This is what is usually used for modeling.  The min value is 1 because a given species in a given year only occurs in the data if it was observed; there are therefore no zeros.  These can be added using the expandBBS_routes() function.}
}}
\source{
\url{www.pwrc.usgs.gov/BBS/RawData}
}
\usage{
BBS_PA
}
\description{
A dataset containing Breeding Bird Survey (BBS) data collected by
the USGS since 1967.  Some columns have been removed to simplify the data.
See ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/Summary.txt for the
original metadata file.
}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding, K. Campbell and M.-A.R. Hudson. 2017. North American Breeding Bird Survey Dataset 1966 - 2016, version 2016.0. U.S. Geological Survey, Patuxent Wildlife Research Center. doi:10.5066/F7W0944J.

Rosenberg et al. 2017.  Use of North American Breeding Bird Survey data in avian conservation assessments
The Condor.  http://www.bioone.org/doi/abs/10.1650/CONDOR-17-57.1
}
\keyword{datasets}
