% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBS_PA_combo_cover_1km.R
\docType{data}
\name{BBS_PA_combo_cover_1km}
\alias{BBS_PA_combo_cover_1km}
\title{Combinations of NLCD land cover classes into more general groups for Pennsylvania, USA.}
\format{A data frame...
\describe{
  \item{Route}{BBS route number}
  \item{BCR}{Bird Conservation Region}
  \item{State}{State}
  \item{route_name}{Route name.  Originally RT..NAME}
  \item{route_length}{length of route.  Originally RT.LENG}
  \item{open_H2O}{NLCD.11}
  \item{developed}{NLCD.21,22,23,24}
  \item{forest}{NLCD.41,42,43}
  \item{grassland}{NLCD.71}
  \item{pasture}{NLCD.81}
  \item{crops}{NLCD.82}
  \item{wetlands}{NLCD.90,95}
  \item{cov_sum}{total landcover accross all variables, not just those in the combined class.  Orginally SUM}
 }}
\source{
\url{https://figshare.com/articles/Data_Paper_Data_Paper/3554424}
}
\usage{
BBS_PA_combo_cover_1km
}
\description{
Landcover within a 1km buffer of Pennsylvania BBS routes.  Original cover
classes combined into overarching groups.  Original data
from the National Land Cover Database (NLCD)
https://catalog.data.gov/dataset/national-land-cover-database-nlcd-land-cover-collection
The data were taken from the 2006 NLCD dataset.
}
\references{
Small et al. 2016.  Data Paper: Local landscape composition and configuration around North American Breeding Bird Survey routes.
See also: http://esapubs.org/archive/ecol/E093/215/
}
\keyword{datasets}
