% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBS_PA_landcover_1km.R
\docType{data}
\name{BBS_PA_landcover_1km}
\alias{BBS_PA_landcover_1km}
\title{Land cover buffers from the 1 km surrounding Breeding Bird Survey Routes (BBS) in Pennsylvania}
\format{A data frame...
\describe{
  \item{Route}{BBS route number}
  \item{BCR}{Bird conservation region}
  \item{State}{State}
  \item{RT..NAME}{Route name}
  \item{RT.LENG}{length of route}
  \item{NLCD.0}{amount of habitat in category 0 of the National Land Cover Database}
  \item{NLCD.11}{NLCD.11}
  \item{NLCD.12}{NLCD.12}
  \item{NLCD.21}{NLCD.21}
  \item{NLCD.22}{NLCD.22}
  \item{NLCD.23}{NLCD.23}
  \item{NLCD.24}{NLCD.24}
  \item{NLCD.31}{NLCD.31}
  \item{NLCD.41}{NLCD.41}
  \item{NLCD.42}{NLCD.42}
  \item{NLCD.43}{NLCD.43}
  \item{NLCD.52}{NLCD.52}
  \item{NLCD.71}{NLCD.71}
  \item{NLCD.81}{NLCD.81}
  \item{NLCD.82}{NLCD.82}
  \item{NLCD.90}{NLCD.90}
  \item{NLCD.95}{NLCD.95}
  \item{sum}{total landcover accross all variables}
 }}
\source{
\url{https://figshare.com/articles/Data_Paper_Data_Paper/3554424}
}
\usage{
BBS_PA_landcover_1km
}
\description{
Landcover within a 1km buffer of Pennsylvania BBS routes.  Original data
from the National Land Cover Database (NLCD)
https://catalog.data.gov/dataset/national-land-cover-database-nlcd-land-cover-collection
The data were taken from the 2006 NLCD dataset.
}
\references{
Small et al. 2016.  Data Paper: Local landscape composition and configuration around North American Breeding Bird Survey routes.
See also: http://esapubs.org/archive/ecol/E093/215/
}
\keyword{datasets}
