% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brook_trout.R
\docType{data}
\name{brook_trout}
\alias{brook_trout}
\title{Brook trout impacts on Chinook salmon survival - original data}
\format{A data frame, including:
\describe{
  \item{site}{Name of study site or stream}
  \item{trout.PRES.ABS}{Are brook trout present in the stream or absent from it?}
  \item{release}{Number of Chinook salmon released}
  \item{recap}{Number of juvenile chinook salmon "recaptured" by detecting their PIT tag; this was not reported in original paper but was estimated using published survival rate.  The survival rate estiamted using mark-recapture methods to account for detection probabilities.}
  \item{surv}{survival rate}
}}
\source{
\url{http://rspb.royalsocietypublishing.org/content/269/1501/1663.short}
}
\usage{
brook_trout
}
\description{
Levin et al. 2002 studied the impacts of brook trout introduced for sport fishing
on native Chinook salmon.  Whitlock & Shulter used a subset
of these data as an example in their textbook Analysis of Biological Data.
}
\examples{

## t-test on survival rate (ignores years)
t.test(surv ~ trout.PRES.ABS, data = brook_trout)

}
\references{
Levin et al.  2002.  Non-indigenous brook trout and the demise of Pacific salmon:
a forgotten threat?  PRSB 269.  DOI: 10.1098/rspb.2002.2063

Whitlock & Schulter.  2nd ed.  Analysis of Biological Data.
}
\keyword{datasets}
