% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_NLCD_classes.R
\name{combine_NLCD_classes}
\alias{combine_NLCD_classes}
\title{Combine multiple landcover classes into a single combined group}
\usage{
combine_NLCD_classes(dat, numeric_cover_codes)
}
\arguments{
\item{dat}{dataframe of NLCD landcover values}

\item{numeric_cover_codes}{numeric codes for landcover classes.  Do not include the "NCLE" prefix.}
}
\description{
This function takes several landcover classes, such as all forest landcover classes
and totals them up.  It is simply wrapper for apply(...sum) that
}
