% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doves_PA.R, R/doves_PA_3_years.R
\docType{data}
\name{doves_PA}
\alias{doves_PA}
\alias{doves_PA}
\title{PA Mourning doves}
\format{A data frame
\describe{
  \item{Year}{Year of study}
  \item{State}{}
  \item{Route}{}
  \item{tot.seen}{}
}}
\usage{
doves_PA

doves_PA
}
\description{
All years of data.  See dataset "doves_PA_3_years" for data that
has already been subset.

...
}
\examples{




}
\references{
Seamans et al 2012.  Mourning dove population status, 2012.

Seamans et al 2012.  Mourning dove population status, 2012.
}
\keyword{datasets}
