% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.R
\docType{data}
\name{eggs}
\alias{eggs}
\title{Bird eggs datatset.}
\format{A data frame with 23 columns, including:
\describe{
  \item{order}{Taxonomic order}
  \item{family}{Taxonomic family}
  \item{spp}{Species}
  \item{asymmetry}{Index of how asymmetric the egg is}
  \item{ellipticity}{Index of how elliptical egg is}
  \item{avgLength.cm}{Average length of egg in cm}

}}
\source{
\url{http://science.sciencemag.org/content/356/6344/1249}
}
\usage{
eggs
}
\description{
A dataset on the shape of eggs.  Original data extract from photographs of multiple eggs.  Values in this dataset are averages.
}
\examples{

## Load packages

library(ggpubr)

ggscatter(data = eggs,
          y = "asymmetry",
          x = "ellipticity")
}
\references{
Stoddard et al.  2017.  Avian egg shape: Form, function, and evolution.
Science.  xxxx:xxx-xxx.
}
\keyword{datasets}
