% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_and_nests.R
\docType{data}
\name{eggs_and_nests}
\alias{eggs_and_nests}
\title{Egg data combined with data on nests}
\format{An object of class \code{data.frame} with 722 rows and 9 columns.}
\usage{
eggs_and_nests
}
\description{
Egg data combined with data on nests
}
\examples{
library(ggplot2)
library(ggpubr)

ggboxplot(data = eggs_and_nests,
y = "ellipticity",
x = "nest")
}
\references{
....
}
\keyword{datasets}
