% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_BBS_routes.R
\name{expand_BBS_routes}
\alias{expand_BBS_routes}
\title{Expand BBS data to include zeros when a species was not seen.}
\usage{
expand_BBS_routes(BBSdata = BBS_PA)
}
\arguments{
\item{BBSdata}{a dataframe of BBS bird counts}
}
\description{
BBS data is distributed by the USGS in a format that only includes data
on species of birds that were actually seen.  If a species was not seen on a route in given
year, it is simply not listed.  That is, instead of zeros to indicate
"that species of bird was not seen" there is nothing.
Also, if a route was not survyed in given year, instead of NA there
is simply no rows of data for that route for that year
This function adds back those empty rows so that zeros occur appropriately
in the data.  HOWEVER, the issue of routes not being done in certain
years has NOT been addressed yet.
}
