% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogarms.R
\docType{data}
\name{frogarms}
\alias{frogarms}
\title{Data on the body size and arm girth of male and female Australian Quacking Frogs (Crinia georgiana)}
\format{A data frame of frog morophology data.
\describe{
  \item{i}{frog Id}
  \item{sex}{Sex of each frog (male or female)}
  \item{mass}{mass of each frog in grams}
  \item{sv.length}{snout-vent length in mm}
  \item{forearm}{forearm girth (?) in mm}
  \item{arm}{arm girth in mm}
 }}
\source{
\url{https://figshare.com/articles/Data_Paper_Data_Paper/3554424}
}
\usage{
frogarms
}
\description{
Morphology information on arm girth and overall body size  was collected by Buzatto et el
as part of a study of sexual selection in the Australian Quacking Frogs.
}
\examples{
## Load packages
library(ggplot2)
library(ggpubr)

## Plot data
### boxplot
ggboxplot(data = frogarms, y = "mass", x = "sex")

### Scatter plot (Figure 1 a of Buzatto et al 2015)
ggscatter(data = frogarms, y = "arm", x = "sv.length",
 color = "sex",
 shape = "sex",
 ylab = "Arm girth (mm)",
 xlab = "Snout-vent length (mm)")
}
\references{
Buzatto BA, Roberts JD, Simmons LW.  2015.  Sperm competition and the evolution of precopulatory weapons: Increasing male density promotes sperm competition and reduces selection on arm strength in a chorusing frog. Evolution 69: 2613-2624. https://doi.org/10.1111/evo.12766

Buzatto BA, Roberts JD, Simmons LW. 2015. Data from: Sperm competition and the evolution of precopulatory weapons: Increasing male density promotes sperm competition and reduces selection on arm strength in a chorusing frog. Dryad Digital Repository. https://doi.org/10.5061/dryad.cr2n4
}
\keyword{datasets}
