% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_heartrate.R
\docType{data}
\name{human_heartrate}
\alias{human_heartrate}
\title{Human heart rate data from American undergraduates}
\format{A data frame with 2 columns:
\describe{
  \item{temp}{body temperature in degrees F}
  \item{sex}{sex of individual}
  \item{heartrate}{heartrate in beats per minute}
}}
\usage{
human_heartrate
}
\description{
Human heart rate data from American undergraduates
}
\examples{

## Plot number of eagles over time
plot(eagles ~ year, data = eaglesPA, col = 2)

}
\keyword{datasets}
