% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_my_data2L.R
\name{make_my_data2L}
\alias{make_my_data2L}
\title{Extract random subset of a dataframe}
\usage{
make_my_data2L(dat, my.code = "nlb24", cat.var, n.sample = 20,
  with.rep = FALSE)
}
\arguments{
\item{dat}{R dataframe to randomly subset}

\item{cat.var}{Categorical variable by which to split the data up}

\item{n.sample}{Number of samples per group to randomly select.  2*n.sample will be the total size of the dataframe.}

\item{with.rep}{Should sampling be with replacement?  If with.rep = TRUE then individuals can occur more than once within the dataframe.}

\item{mycode}{Unique code, such as your email address.  Must contain only letters and numbers and not the at symbol or other punctuation.}
}
\description{
Currently only works with data with 2-levels.  Notes that if the size of the dataframe is too small to accomodate n.sample then with.rep is set to TRUE.
}
\examples{
#Subset frogarms data
data(frogarms)
my.frogs <- make_my_data2L(dat = frogarms,
                           my.code = "nlb24",
                           cat.var = "sex",
                           n.sample = 20,
                           with.rep = FALSE)

summary(my.frogs)

}
