% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_to_zero.R
\name{NA_to_zero}
\alias{NA_to_zero}
\title{Change all NA values (missing vlaues) in a column to zero}
\usage{
NA_to_zero(dat, column, return = "df")
}
\arguments{
\item{dat}{dataframe of USGS BBS data containing NAs that should be replaced with zeros}

\item{column}{column within dataframe containing NAs that should be converted to a dataframe}

\item{return}{"df" or "column"; whether to return the whole dataframe (deafault) or just the column}
}
\description{
Change all NA values (missing vlaues) in a column to zero
}
