% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_test_ES.R
\name{plot_t_test_ES}
\alias{plot_t_test_ES}
\title{Function to plot the absolute effect size from a t-test}
\usage{
plot_t_test_ES(t.test.out, point.sz = 5)
}
\arguments{
\item{t.test.out}{Output of 2-sample t-test using t.test()}

\item{point.sze}{Size of point representing the mean.}
}
\description{
This function takes a fitted t-test as its input and plots a graph of the absolute effect size (difference between means).
Note that this is just the raw difference between means of each group, not a standardized effect size.
Data that is plotted is extracted from the means and confidence intervals reported by the t.test() function.
Uses ggplot
}
\examples{
#Subset frogarms data
data(frogarms)
mass.t <- t.test(mass ~ sex, data = frogarms)
plot_t_test_ES(mass.t)

}
