% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTukeysHSD.R
\name{plotTukeysHSD}
\alias{plotTukeysHSD}
\title{Plot output of TukeyHSD() function when there are 3 groups}
\usage{
plotTukeysHSD(tukey.out, x.axis.label = "Comparison",
  y.axis.label = "Effect Size", axis.adjust = 0)
}
\arguments{
\item{tukey.out}{output of TukeyHSD()}

\item{x.axis.label}{what should be plotted on the x axis}

\item{y.axis.label}{what should be plotted on the y axis}

\item{axis.adjust}{for tweaking plotting}
}
\description{
Plot output of TukeyHSD() function when there are 3 groups
}
