% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dove_routes.R
\name{sample_dove_routes}
\alias{sample_dove_routes}
\title{Select random subset of Breeding Bird Survey (BBS) routes}
\usage{
sample_dove_routes(dat = doves_PA, years = c(1990, 2000, 2005),
  max.sample = 50)
}
\arguments{
\item{dat}{Dataframe containing BBS data.  Must contain columns for Year, Aou species code and Route}

\item{years}{Years for for which you want to generate data for}

\item{max.sample}{Maximum number routes per year.}
}
\description{
This function randomly samples routes from within a stage.
}
\examples{

## create data
#dat <- sample_dove_routes(dat = doves_PA,
 #       years = c(2005,2007,2010))

## Make year a factor
#dat$Year <- factor(dat$Year)

## Plot means with 95\% confidence intervals

library(ggplot2)
library(ggpubr)

#ggerrorplot(dat,
 # x = "Year",
  #y = "tot.seen",
  #desc_stat = "mean_ci",
  #add = "mean")

}
