% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trillium.R
\docType{data}
\name{trillium}
\alias{trillium}
\title{Data on the size of Trillium from several sites in Crawford County, PA}
\format{A data frame
\describe{
  \item{site}{sites code}
  \item{spp}{species. All are = T for Trillium}
  \item{length.stem.cm}{length from soil to leaf whorl}
  \item{length.leaf.cm}{length of longest leaf in cm}
}}
\usage{
trillium
}
\description{
The size of Trillium has been hypothesized to be correlated with
the intensity of deer browse occuring on thee population.
}
\references{
Author year.
}
\keyword{datasets}
