% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_NLCD_classes.R
\name{combine_NLCD_classes}
\alias{combine_NLCD_classes}
\title{Combine multiple landcover classes into a single combined group}
\usage{
combine_NLCD_classes(dat, numeric_cover_codes)
}
\arguments{
\item{dat}{dataframe of NLCD landcover values}

\item{numeric_cover_codes}{numeric codes for landcover classes}
}
\description{
Combine multiple landcover classes into a single combined group
}
