% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs_and_nests.R
\docType{data}
\name{eggs_and_nests}
\alias{eggs_and_nests}
\title{Egg data combined with data on nests}
\format{A data frame
\describe{
  \item{family}{Taxonomic family}
  \item{order}{Taxonomic order}
  \item{spp}{Genus and species}
  \item{asymmetry}{Index of how asymmetric an egg is}
  \item{ellipticity}{Index of how elliptical an egg is}
  \item{ave.length.com}{Mean length of egg}
  \item{nest}{nest type of species}
  \item{source}{author}
  \item{bird.mass}{mass of bird in grams}
}}
\usage{
eggs_and_nests
}
\description{
Egg data combined with data on nests
}
\examples{
library(ggplot2)
library(ggpubr)

ggboxplot(data = eggs_and_nests,
y = "ellipticity",
x = "nest")
}
\references{
....
}
\keyword{datasets}
