% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_BBS_routes.R
\name{sample_BBS_routes}
\alias{sample_BBS_routes}
\title{Select random subset of Breeding Bird Survey (BBS) routes}
\usage{
sample_BBS_routes(dat = BBS_PA, years = c(1990, 2000, 2005),
  aou.code = 7610, max.sample = 50)
}
\arguments{
\item{dat}{Dataframe containing BBS data.  Must contain columns for Year, Aou species code and Route}

\item{years}{Years for for which you want to generate data for}

\item{aou.code}{American Ornithological Union (AOU) numeric species code}

\item{max.sample}{Maximum number routes per year.}
}
\description{
This function randomly samples routes from within a stage.
}
